Set PATH outside of the shell invocation
I was doing this inside the shell previously to make sure shy would get
in the last word, in case a bashrc file or something would overwrite the
PATH again and set rm/mv/cp back.

Problem is that if the shell command is invalid, the path setting ends
up in the error messages. That's confusing. So for now let's set the
PATH on the outside of the shell, and figure out other approaches if it
turns out to be needed.